requirements.txt
Pillow
pycryptodome
stegano